Working on the Rooms, SearchView is fixed, refactor Command because we are not using it otherwise.

+++ src/main/java/ie/ul/hbs2/booking/BookingManager.java
@@ -63,7 +63,7 @@ public class BookingManager implements IPaymentCallback {
         int memLVL = book.getMemLvl();
         double totalSpent = 0 ;
         BookingManager temp = new BookingManager();
-        int roomCost =  getRoomCost(book.getRoomType());
+        double roomCost =  getRoomCost(book.getRoomType());
         float discount;
         discount = RewardFactory.getReward(memLVL).get_discount() / 100;
         totalSpent = ((roomCost * nights) * rmBooked);
